# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gotion API
  version: 1.0.0
servers:
  - url: 'http://localhost:8000'
    description: 開発環境
paths:
...
  '/oauth/notion/callback':
    get:
      description: Notion OAuth認証
      operationId: NotionOAuth
      parameters:
        - name: code
          description: 認証コード
          in: path
          schema:
            type: string
          required: true
        - name: stateFromRequest
          description: CSRF対策用のstate Queryから取り出し
          in: path
          schema:
            type: string
          required: true
        - name: stateFromCookie
          description: CSRF対策用のstate Cookieから取り出し
          in: cookie
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect Gotion DB Name Page (/step/input-db-name)
          headers:
            Set-Cookie:
              description: notionのaccess tokenを暗号化
              schema:
                type: string
                example: name=notion-access-token; Path=/; Max-Age=24*60; Secure; HttpOnly; SameSite=None;
...